Fixed tests, added static goTo method to App
