+++ js/intro/intro.js
@@ -1,5 +1,5 @@
-import {views} from '../data/gamedata';
-import routes from '../main';
+import {views, ControllerID} from '../data/gamedata';
+import App from '../main';
 import IntroView from './intro-view';
 import insertTemplate from '../utils/insert-template';
 import ScreenPresenter from '../screen';
@@ -17,7 +17,7 @@ export default class IntroScreen extends ScreenPresenter {
 
     this._view.onStartClick = () => {
       gameState.changeState({template: views.greeting});
-      location.hash = routes.GREETING;
+      App.goTo(ControllerID.GREETING);
     };
   }
 }
