+++ js/main.js
@@ -1,3 +1,4 @@
+import {ControllerID} from './data/gamedata';
 import IntroScreen from './intro/intro';
 import GreetingScreen from './greeting/greeting';
 import RulesScreen from './rules/rules';
@@ -5,18 +6,10 @@ import GameScreen from './game/game';
 import StatsScreen from './stats/stats';
 
 
-const ControllerID = {
-  INTRO: ``,
-  GREETING: `greeting`,
-  RULES: `rules`,
-  GAME: `game`,
-  STATS: `stats`,
-};
-
 const getControllerIDFromHash = (hash) => hash.replace(`#`, ``);
 
 
-class Application {
+export default class Application {
   constructor() {
     this._routes = {
       [ControllerID.INTRO]: IntroScreen,
@@ -26,9 +19,9 @@ class Application {
       [ControllerID.STATS]: StatsScreen,
     };
 
-    window.onhashchange = () => {
+    addEventListener(`hashchange`, () => {
       this._changeController(getControllerIDFromHash(location.hash));
-    };
+    }, false);
   }
 
   init() {
@@ -45,8 +38,10 @@ class Application {
       throw new Error(`Invalid router: location.hash must be an entry of Application._routes.`);
     }
   }
+
+  static goTo(route) {
+    location.hash = route;
+  }
 }
 
 new Application().init();
-
-export default ControllerID;
