+++ package.json
@@ -32,8 +32,6 @@
     "gulp-sass": "3.1.0",
     "gulp-sourcemaps": "2.6.0",
     "gulp-uglify": "3.0.0",
-    "jsdom": "11.0.0",
-    "jsdom-global": "3.0.2",
     "rollup-plugin-babel": "2.7.1"
   },
   "engines": {
